refactor: Merge hardware/radio config pairs — consolidate device data, remove dead code
Consolidate HARDWARE_DEVICES dict and HardwareDevice dataclass from
hardware_config.py into hardware.py (single source of truth for device data).

Delete 3 vestigial files:
- hardware_config.py: HardwareConfigurator was unused (TUI uses dialog-based
  meshtasticd_config_mixin instead)
- radio.py: RadioConfigurator never imported anywhere
- radio_config.py: RadioConfig/radio_config_menu() never called (TUI handles
  radio config natively)

Net reduction: ~1,695 lines. All 2,123 tests pass.
